feat: added function to sample uint32_t in distribution_sampler.h

+++ src/multinomial_sampler.cpp
@@ -4,11 +4,11 @@ int multinomial_sampler::sample() {
   return this->dist(this->rng);
 }
 
-void multinomial_sampler::sample(std::vector<double> &out, uint32_t n) {
+void multinomial_sampler::sample(std::vector<uint32_t> &out, uint32_t n) {
   out.clear();
   out.reserve(n);
   for (uint32_t i = 0; i < n; ++i) {
-    out.push_back(static_cast<double>(this->dist(this->rng)));
+    out.push_back(this->dist(this->rng));
   }
 }
